+++ src/main/java/net/helix/pendulum/service/validatomanager/CandidateTracker.java
@@ -1,4 +1,4 @@
-package net.helix.pendulum.service.curator;
+package net.helix.pendulum.service.validatomanager;
 
 import net.helix.pendulum.controllers.TransactionViewModel;
 import net.helix.pendulum.model.Hash;
@@ -10,9 +10,9 @@ public interface CandidateTracker {
     int getStartRound();
 
     /**
-     * curator address.<br />
+     * validatomanager address.<br />
      */
-    //Hash curatorAddress = HashFactory.ADDRESS.create("c2eb2d5297f4e70f3e40e3d7aa3f5c1d7405264aeb72232d06776605d8b61211");
+    //Hash validatorManagerAddress = HashFactory.ADDRESS.create("c2eb2d5297f4e70f3e40e3d7aa3f5c1d7405264aeb72232d06776605d8b61211");
 
     /**
      * Analyzes the given transaction to determine if it is a valid candidate application.<br />
@@ -21,9 +21,9 @@ public interface CandidateTracker {
      *
      * @param transaction the transaction that shall be examined
      * @return {@code true} if the milestone could be processed and {@code false} if the bundle is not complete, yet
-     * @throws CuratorException if anything unexpected happens while trying to analyze the milestone candidate
+     * @throws ValidatorManagerException if anything unexpected happens while trying to analyze the milestone candidate
      */
-    boolean processCandidate(TransactionViewModel transaction) throws CuratorException;
+    boolean processCandidate(TransactionViewModel transaction) throws ValidatorManagerException;
 
     /**
      * Does the same as {@link #processCandidate(TransactionViewModel)} but automatically retrieves the
@@ -31,14 +31,14 @@ public interface CandidateTracker {
      *
      * @param transactionHash the hash of the transaction that shall be examined
      * @return {@code true} if the milestone could be processed and {@code false} if the bundle is not complete, yet
-     * @throws CuratorException if anything unexpected happens while trying to analyze the milestone candidate
+     * @throws ValidatorManagerException if anything unexpected happens while trying to analyze the milestone candidate
      */
-    boolean processCandidate(Hash transactionHash) throws CuratorException;
+    boolean processCandidate(Hash transactionHash) throws ValidatorManagerException;
 
 
-    Set<Hash> getNominees();
+    Set<Hash> getValidators();
 
-    Set<Hash> getNomineesOfRound(int roundIndex);
+    Set<Hash> getValidatorsOfRound(int roundIndex);
 
     /**
      * This method starts the background worker that...